+++ test/features/steps/basic_commands.py
@@ -73,9 +73,30 @@ def step_see_found(context):
         ''') + context.conf['pager_boundary'],
         timeout=5
     )
+
+
 @then(u'we confirm the destructive warning')
 def step_confirm_destructive_command(context):
     """Confirm destructive command."""
     wrappers.expect_exact(
         context, 'You\'re about to run a destructive command.\r\nDo you want to proceed? (y/n):', timeout=2)
     context.cli.sendline('y')
+
+
+@when(u'we answer the destructive warning with "{confirmation}"')
+def step_confirm_destructive_command(context, confirmation):
+    """Confirm destructive command."""
+    wrappers.expect_exact(
+        context, 'You\'re about to run a destructive command.\r\nDo you want to proceed? (y/n):', timeout=2)
+    context.cli.sendline(confirmation)
+
+
+@then(u'we answer the destructive warning with invalid "{confirmation}" and see text "{text}"')
+def step_confirm_destructive_command(context, confirmation, text):
+    """Confirm destructive command."""
+    wrappers.expect_exact(
+        context, 'You\'re about to run a destructive command.\r\nDo you want to proceed? (y/n):', timeout=2)
+    context.cli.sendline(confirmation)
+    wrappers.expect_exact(context, text, timeout=2)
+    # we must exit the Click loop, or the feature will hang
+    context.cli.sendline('n')