+++ test/test_special_iocommands.py
@@ -1,4 +1,6 @@
 # coding: utf-8
+from __future__ import unicode_literals
+
 import os
 import stat
 import tempfile
@@ -232,3 +234,42 @@ def test_watch_query_interval_clear(clear_mock):
                                  cur=cur))
         test_asserts(watch_query('-c {0!s} select 1;'.format(seconds),
                                  cur=cur))
+
+
+def test_split_sql_by_delimiter():
+    for delimiter_str in (';', '$', '😀'):
+        mycli.packages.special.set_delimiter(delimiter_str)
+        sql_input = "select 1{} select \ufffc2".format(delimiter_str)
+        queries = (
+            "select 1",
+            "select \ufffc2"
+        )
+        for query, parsed_query in zip(
+                queries, mycli.packages.special.split_queries(sql_input)):
+            assert(query == parsed_query)
+
+
+def test_switch_delimiter_within_query():
+    mycli.packages.special.set_delimiter(';')
+    sql_input = "select 1; delimiter $$ select 2 $$ select 3 $$"
+    queries = (
+        "select 1",
+        "delimiter $$ select 2 $$ select 3 $$",
+        "select 2",
+        "select 3"
+    )
+    for query, parsed_query in zip(
+            queries,
+            mycli.packages.special.split_queries(sql_input)):
+        assert(query == parsed_query)
+
+
+def test_set_delimiter():
+
+    for delim in ('foo', 'bar'):
+        mycli.packages.special.set_delimiter(delim)
+        assert mycli.packages.special.get_current_delimiter() == delim
+
+
+def teardown_function():
+    mycli.packages.special.set_delimiter(';')